Fix: split pthread_port into logical sections
