+++ deploy/packaging/debian/devel.noarch
@@ -30,10 +30,12 @@
 ./usr/local/mdsplus/include/camshr.h
 ./usr/local/mdsplus/include/camshr_messages.h
 ./usr/local/mdsplus/include/classdef.h
+./usr/local/mdsplus/include/condition.h
 ./usr/local/mdsplus/include/coz.h
 ./usr/local/mdsplus/include/dbidef.h
 ./usr/local/mdsplus/include/dcl.h
 ./usr/local/mdsplus/include/dtypedef.h
+./usr/local/mdsplus/include/getusername.h
 ./usr/local/mdsplus/include/int128.h
 ./usr/local/mdsplus/include/ipdesc.h
 ./usr/local/mdsplus/include/libroutines.h
@@ -66,6 +68,7 @@
 ./usr/local/mdsplus/include/rtevents.h
 ./usr/local/mdsplus/include/servershr.h
 ./usr/local/mdsplus/include/servershr_messages.h
+./usr/local/mdsplus/include/socket_port.h
 ./usr/local/mdsplus/include/sqldb.h
 ./usr/local/mdsplus/include/sqlfront.h
 ./usr/local/mdsplus/include/status.h
