+++ mdstcpip/RemoveConnection.c
@@ -22,9 +22,10 @@ CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY,
 OR TORT (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE
 OF THIS SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 */
-#include "mdsip_connections.h"
 #include <stdlib.h>
+
 #include <treeshr.h>
+#include "mdsip_connections.h"
 
 extern int TdiSaveContext();
 extern int TdiDeleteContext();
