+++ mdstcpip/ioroutinesx.h
@@ -16,14 +16,8 @@ static IoRoutines io_routines = {
   io_connect, io_send, io_recv, io_flush, io_listen, io_authorize, io_reuseCheck, io_disconnect, io_recv_to, io_check
 };
 #include <mdsshr.h>
+
 #include <inttypes.h>
-#ifdef _TCP
- #ifdef _WIN32
-  #define close closesocket
- #endif
-#else
- #include <poll.h>
-#endif
 
 #define IP(addr)   ((uint8_t*)&addr)
 #define ADDR2IP(a) IP(a)[0],IP(a)[1],IP(a)[2],IP(a)[3]
